Renamed newDocs/ to music21/documentation.
The directory contains Python code, and placing it into the music21 package
hierarchy allows that code to be doctested, unittested, and -importantly-
imported and examined by any other code in the music21 or wider Python
packagesystem.

So, not only can the code in documentation/ be doctested without any special
casing, but the code in documentation can itself be documented just like any
of the other code in music21.