Implemented a module/package ignore mechanism in the new doc building process.
Add the line

    _DOC_IGNORE_MODULE_OR_PACKAGE = True

to any python file in music21 to force the doc building script to ignore it.

If the incantation is placed in an __init__.py file, all other modules and
package in that entire directory will also be ignored.

The following files now contain the incantation:

    music21/abc/testFiles.py
    music21/analysis/phrasing.py
    music21/chordTables.py
    music21/classCache.py
    music21/configure.py
    music21/humdrum/testFiles.py
    music21/musedata/testFiles.py
    music21/musicxml/testFiles.py
    music21/romanText/testFiles.py
    music21/test/__init__.py
    music21/xmlnode.py

Note that all music/test/*.py files are no longer documented.  We can put them
back if we feel like it, simply by deleting the line in test/__init__.py.  This
is really just proof of concept as far as that goes.

Some files are still manually ignored by the documentation.ModuleIterator
class:

    music21/webapps/base-archive.py
    music21/trecento/exceldiff.py

These are ignored by hand because they cannot be imported and inspected for
the ignore incantation without causing errors.