Fix a few failing tests in music21.braille.test.
Some were missing the "test." prefix when calling the example function.
Others use .pop(0) to remove the time signature after makeNotation,
which does not do what the code expects.  It actually pops the clef.
Since the time signature is essential for reading braille music anyway,
stop trying to remove the time signature and include it in the example output.

This fixes 13 out of 48 failing tests.