WIP-MEI: Create an 'MeiToM21Converter' class, as per Myke's suggestion
So far, the class is merely a wrapper for mei.base.convertFromString(). The
plan is that the class will hold shared state among all the converer functions,
avoiding the need to pass around things like the "slurBundle," for example. It
may also help in the effort to simplify convertFromString(), which will
eventually disappear into this class anyway.