WIP-MEI: Remove "star" import to "mei" module
Previously, everything in the music21.mei.base file was imported to the
music21.mei module at runtime. Now, only the MeiToM21Converter object is
imported. This should help clarify to users that they probably don't
want to be using the functions in the "base" module unless they're really
doing programming with MEI.