WIP-MEI: Add complicated test file for MEI import
It's quite ridiculous indeed! But you can use it to inspect the music21 objects,
and see that the MEI importer converters everything correctly. You can also try
exporting the imported score to MusicXML and visualize import quality that way,
but several weird things won't be exported properly.

There's also a LilyPond score and PDF output of the intended musical meaning!