MEI: Fix: Do not call padAsAnacrusis() on the first Measure
This was a hold-over from when Part objects were assembled in a different way
(by repeatedly calling append() on the Part as each Measure was converted). Now
that we wait until all Measure objects are completed before calling append(),
this call to padAsAnacrusis() is superfluous, so I took it out.