Fixed: MEI transposition information of instruments not parsed
Previously, the mei-to-m21 converter produced instrument metadata in a way
that other music21 modules weren't expecting. That was fixed in a previous
commit.

Additionally, toSoundingPitch() failed because we didn't set the
"atSoundingPitch" attribute while importing. MEI is always notated at written
pitch, so this is now always set to False.

Closes #125