Use subprocess.run instead of os.system for PNG generation
On Windows 10 inside Jupyter Notebook environment show method with PNG output does not work because of the double quotes placed around the arguments of the called command. Replacing the call for PNG conversion done via `os.system` with `subprocess` fixes the  problem. This problem and its solution were mentioned in issue #348