Update instrument.py
1. Introduces and implements the following variables (as explained in the expanded docs):
    * hardLimitOnLowest
    * hardLimitOnHighest
    * meanPitch

2. Adds Contrabassoon

3. Transpositions. This warrants a small essay ;)

Most instruments seem to have their lowest note given in written pitch along with the (most typical) transposition interval.

For instance,

Clarinet:
        self.lowestNote = pitch.Pitch('E3')
        self.transposition = interval.Interval('M-2')
The transposition indicates that we're working with a Bb clarinet by default, in which case the lowest pitch is certainly the written pitch (sounding D3).

Likewise the Bass Clarinet:
        self.lowestNote = pitch.Pitch('E-3')
        self.transposition = interval.Interval('M-9')
Again, with this (standard) transposition, this lowest pitch most likely indicated written pitch (sounding D-2).

I'm not altogether convinced that this is the best system, especially when we get to the double bass, for which the doc comments read:
'For the Contrabass (or double bass), the stringPitches attribute refers to the sounding pitches of each string; whereas the lowestNote attribute refers to the lowest written note.'

All the same, people may be relying on the defaults as they stands, so I've just made the minimal change for consistency and accuracy, e.g.:
- the 'E3' lowest for the English Horn needed changing to 'B3'.
- the 'C5' lowest for the Piccolo needed moving down an octave (and changing to 'D4' while we're at it ... ).
(It was also possible to promote the lowest pitch for saxophones to the super class).

In future, we could consider a clarification that distinguishes between self.lowestWritten and self.lowestSounding.
Probably we'd set self.lowestNote = self.lowestSounding for continuity, at least for the time being.

Fortunately, rumour has it that there's a clarinetist somewhere high up the music21 ranks ... that person will be ideally placed to adjudicate on all this ;)