Periods in instrument abbreviations
Strictly, these should be there for all cases of incomplete words (e.g. 'b.' for 'bass') and not otherwise (e.g. no period in 'bsn' for 'bassoon') but music21 removes punctuation before evaluating (see instrument.fromString).

This means that, nothing is gained by having all the duplication of abbreviations with and without periods in instrumentLookup, like
    'acc': 'accordion',
    'acc.': 'accordion',

Moreover, the those periods were being replaced with nothing, rather than a space, meaning that many legitimate abbreviations fail.

To fix this, I've added a line replacing periods specifically with spaces, before running removePunctuation.

In any case, the duplicates can go, so I've gone ahead and removed them all.

In future (TODO) we need better handling of this with a regex that can parse typical structures like '{instrument} in {key}' for transposing instruments. For instance, the abbreviation 'a' for 'alto' cannot be used as the code is currently set up because of conflict with the 'a' in 'Clarinet in A', for instance. This commit is still an improvement in the meantime, because the previous 'a.' for 'alto' abbreviation wasn't being used anyway ('a.' simply couldn't be called due to the removal of punctuation from all input strings) but it is only part of the problem.

I'll create an issue.