Change instrumentLookup to use simple names (#631)
* Update instrumentLookup.py

Makes allToBestName from the consituent dicts to ensure consistency. Adds test for the same.

* Update instrumentLookup.py

Move import statement to top;
Empty line lint.

* Update instrumentLookup.py

Adds bass trombone;
Removes unused 'bestNameToLanguage' dict

* Update instrumentLookup.py

Iterate the dictionary directly

* Update instrument.py

saxofono tenor[e]

* Update instrumentLookup.py

Saxonofone types;
Special case of transliteration via the relevant language

* Update instrumentLookup.py

Reinstate a few cases that were in the former allToBestName and not in any of the constituent dicts (a good sign that building from the combined form is needed for consistency.)
TODO: more systemtic handling e.g. of plurals

* Fix lints and make English names authoritative

* Update instrument.py

1. Introduces and implements the following variables (as explained in the expanded docs):
    * hardLimitOnLowest
    * hardLimitOnHighest
    * meanPitch

2. Adds Contrabassoon

3. Transpositions. This warrants a small essay ;)

Most instruments seem to have their lowest note given in written pitch along with the (most typical) transposition interval.

For instance,

Clarinet:
        self.lowestNote = pitch.Pitch('E3')
        self.transposition = interval.Interval('M-2')
The transposition indicates that we're working with a Bb clarinet by default, in which case the lowest pitch is certainly the written pitch (sounding D3).

Likewise the Bass Clarinet:
        self.lowestNote = pitch.Pitch('E-3')
        self.transposition = interval.Interval('M-9')
Again, with this (standard) transposition, this lowest pitch most likely indicated written pitch (sounding D-2).

I'm not altogether convinced that this is the best system, especially when we get to the double bass, for which the doc comments read:
'For the Contrabass (or double bass), the stringPitches attribute refers to the sounding pitches of each string; whereas the lowestNote attribute refers to the lowest written note.'

All the same, people may be relying on the defaults as they stands, so I've just made the minimal change for consistency and accuracy, e.g.:
- the 'E3' lowest for the English Horn needed changing to 'B3'.
- the 'C5' lowest for the Piccolo needed moving down an octave (and changing to 'D4' while we're at it ... ).
(It was also possible to promote the lowest pitch for saxophones to the super class).

In future, we could consider a clarification that distinguishes between self.lowestWritten and self.lowestSounding.
Probably we'd set self.lowestNote = self.lowestSounding for continuity, at least for the time being.

Fortunately, rumour has it that there's a clarinetist somewhere high up the music21 ranks ... that person will be ideally placed to adjudicate on all this ;)

* Periods in instrument abbreviations

Strictly, these should be there for all cases of incomplete words (e.g. 'b.' for 'bass') and not otherwise (e.g. no period in 'bsn' for 'bassoon') but music21 removes punctuation before evaluating (see instrument.fromString).

This means that, nothing is gained by having all the duplication of abbreviations with and without periods in instrumentLookup, like
    'acc': 'accordion',
    'acc.': 'accordion',

Moreover, the those periods were being replaced with nothing, rather than a space, meaning that many legitimate abbreviations fail.

To fix this, I've added a line replacing periods specifically with spaces, before running removePunctuation.

In any case, the duplicates can go, so I've gone ahead and removed them all.

In future (TODO) we need better handling of this with a regex that can parse typical structures like '{instrument} in {key}' for transposing instruments. For instance, the abbreviation 'a' for 'alto' cannot be used as the code is currently set up because of conflict with the 'a' in 'Clarinet in A', for instance. This commit is still an improvement in the meantime, because the previous 'a.' for 'alto' abbreviation wasn't being used anyway ('a.' simply couldn't be called due to the removal of punctuation from all input strings) but it is only part of the problem.

I'll create an issue.

* Update instrument.py

Removes proposed new attributes for ‘meanPitch’ (pitch) and hardLimit (boolean).

So the net changes on this PR involve:
- Consistency in written / sounding pitch (instrument.py)
- Adding contrabassoon, bass trombone (both)
- Improvements to the fromString method (instrument.py) especially …
- Handling of abbreviations (with / without periods), plurals, languages etc.(instrumentLookup.py) ,

* Fix typo – invoke the right test case ;)

* One more abbreviation for @jacobtylerwalls