Adds hasRepeatedChordStep(1) check in addition to 3 and 5 for assessing chord quality
Extends the check for chords containing more than one triad.

Previously hasRepeatedChordStep() checked for 3 and 5, so that:
>>> chord.Chord('C E- E G').quality
'other'
>>> chord.Chord('C E G- G').quality
'other'

But that still returned 'major' for the following case:
>>> chord.Chord('C C# E G').quality
'major'

While it's true that there's a major triad in there (CEG), there's also a diminished one (C#EG); it would be more consistent if this cases returns 'other' too.

With this PR, that case now returns 'other'
>>> chord.Chord('C C# E G').quality
'other'