Composite Lyric objects (#794)
* Composite Lyric objects

Changes #673 and #666 to deal with multiple syllables in a single verse in a different way:

Lyric objects gain a .components[] list which is a list of subservient note.Lyric() objects representing multiple lyrics in a single verse (like "co_e" in "Il bianco e dolce cigno" or "th'an" in Hark the Herald Angels sing, "With th'angelic host proclaim")  -- in these cases a `.isComposite` boolean is true, and the syllabic is set to "composite" -- roundtrips to musicxml.

I appreciate @DIDONEproject work on putting a list in the .text and .syllabic attributes, but this ended up other code in real-world scenarios, so a new direction needed to be taken.  DIDONE and others who rely on the direction used since November 2020 will need to refactor.  Though if you only want the .text attribute after parsing, it should work fine.

* fix and test searching of multiple lyrics in a part

Both multiple verses and composite Lyrics in a part.

* re.Pattern does not exist on 3.6