Sensible Defaults for Augmented 6th
Fixes #529

Diagnoses that there are larger problems with postFigureFromChordAndKey  not taking into account major and minor differences.  It needs a major refactor  BUT this is a step in the right direction and for helping give many better test cases

Also fixes a small bug in Mordents and intervals in Expressions and adds some typing in makeNotation -- it's not related but I accidentally had it sitting around the same directory.