Stop storing id(x) as .id
The `.id` of the element allows for having the memory location only as a crutch so that it's not required.  It should not be stored but generated dynamically.

This was a mistake in architecture created before music21 v1 release and it's finally time to fix.

TODO: fix in beam and tie also.