Restore original instrument lengths after _transposeByInstrument
By not restoring original instrument lengths we were creating instruments with quarter lengths like 40.0, which is a complex duration that was being split by  `splitAtDurations`.  Then we'd end up with two Instrument objects in measure 1, one of which was FAR FAR later, so even though it was being removed in output, it caused a massive `<forward>` tag in the musicxml output which caused spacing to be terrible.