optimize opFrac
see post to music21list about details.

Switches the cache from 1024 to None -- not because we want to cache more than 1024 values, but to avoid the overhead of an LRU cache, which is significant.

Speeds up opFrac(0.5) by about 37% at a cost of adding 25% to opFrac(9/16) or 5% to opFrac(1/3) -- worth it -- though the lru_cache speedup makes the latter actually faster.