Emphasize Pitches in Interval classes (#1396)
Store only Pitch elements in Interval(), not Notes, and depend on Pitch._client to get the notes back.

Simplifies the Interval object and speeds it up if we're not constantly looking at Pitch or Note at any moment. Improves typing.

More importantly, the client for the Pitch will not change when transposed inPlace and therefore chordAttached on the note will also not change. This was important with cached methods like isConsonant when one of the individual pitches in the Chord was being transposed. Now client is informed.