Clarify entities
Clarify copyright status of music21 -- license is still BSD-3-clause -- Open Source -- no worries!

There was never a legal entity called "the music21 project" nor "cuthbertLab" -- both informal names; the legal copyright needs to rest in a legal entity, I've been advised to clarify what that is.  Hopefully that legal entity can be created in the future, but for now and for legal clarity, it is more precise to describe MSAC as the sole copyright holder.  But he can't restrict the license of anything released (nor does he plan to!).  It's been bugging me since v1 and always a good time to fix before a release.