Do not transpose altered 7,9,11,13 in bII7
According to published usage, transposed chords (bII, etc.) should take their 7th from the key and not be transposed.  So in c-minor, bII7 should be D-, F, A-, C, and not with C-flat.

Assuming that this would apply to 9ths, 11ths, and 13ths as well.

Does not affect interpretation of vi7, VI7, etc., I don't believe

Fixes #1410

Credit MalcolmSailor for the notice and research on this.