Fix stripTies when accidentals are neutral & none
When a stream contains two notes/chords that contain the same pitch, but one has a "neutral" accidental and the other one doesn't have an accidental at all (None), the tied notes weren't matched even though they should.

It's debatable whether this change should really take place in the pitch.Pitch class, where the `__eq__` method could be adapted to interpret a "neutral" accidental to be equal to a None accidental. 
This seemed like the smaller change, but I'm happy to change the PR.
Let me know what you think!