Speed up quantize()
While iterating over all streams, rests of duration 0 are deleted. Previously, the list of rests to be deleted was not reset after finishing with a substream, yielding extremely slow performance on big scores as the code attempted to delete all previously found (and already deleted) rests again in each following substream.