SoundParser to a separate file
Going in the opposite direction as #1599 -- move MusicXML `<sound>` tag parsing to a separate file along with several helpers to helpers.py -- both m21ToXML and xmlToM21 are getting huge! so anything to try to make these files more manageable is the right direction.

While fixing mypy errors, noticed a bunch of warnings that things that were typed were not being checked in noteworthy/binaryTranslate.py -- checked these and fixed.  (TODO: make the contents of untyped functions are not checked become an error in CI)