Python 3.12 Compatibility
Make music21 test suite fully compatible with Python 3.12 -- because the basic functionality of music21 works without change in music21, I'm not going to break sabbatical to do a new release for this, but all code tests going forward should also run on 3.12.

Minimum version remains 3.10

Thanks to @jacobtylerwalls for the fix in advance for the new OrderedDict syntax.